feat: add a generator for shell completions
