+++ crates/aiken/Cargo.toml
@@ -31,6 +31,7 @@ aiken-lang = { path = "../aiken-lang", version = "1.0.13-alpha" }
 aiken-lsp = { path = "../aiken-lsp", version = "1.0.13-alpha" }
 aiken-project = { path = '../aiken-project', version = "1.0.13-alpha" }
 uplc = { path = '../uplc', version = "1.0.13-alpha" }
+clap_complete = "4.3.2"
 
 [build-dependencies]
 built = { version = "0.6.0", features = ["git2"] }
