+++ crates/aiken/src/main.rs
@@ -1,49 +1,14 @@
 use aiken::{
-    built_info,
     cmd::{
+        Cmd,
         blueprint::{self, address},
         build, check, docs, fmt, lsp, new,
         packages::{self, add},
-        tx, uplc,
+        tx, uplc, completion,
     },
 };
-use clap::Parser;
+
 use owo_colors::OwoColorize;
-
-/// Aiken: a smart-contract language and toolchain for Cardano
-#[derive(Parser)]
-#[clap(version = version(), about, long_about = None)]
-#[clap(propagate_version = true)]
-pub enum Cmd {
-    New(new::Args),
-    Fmt(fmt::Args),
-    Build(build::Args),
-    Address(address::Args),
-    Check(check::Args),
-    Docs(docs::Args),
-    Add(add::Args),
-
-    #[clap(subcommand)]
-    Blueprint(blueprint::Cmd),
-
-    #[clap(subcommand)]
-    Packages(packages::Cmd),
-
-    #[clap(subcommand)]
-    Tx(tx::Cmd),
-
-    #[clap(subcommand)]
-    Uplc(uplc::Cmd),
-
-    #[clap(hide = true)]
-    Lsp(lsp::Args),
-}
-
-impl Default for Cmd {
-    fn default() -> Self {
-        Self::parse()
-    }
-}
 
 fn main() -> miette::Result<()> {
     panic_handler();
@@ -61,15 +26,8 @@ fn main() -> miette::Result<()> {
         Cmd::Lsp(args) => lsp::exec(args),
         Cmd::Tx(sub_cmd) => tx::exec(sub_cmd),
         Cmd::Uplc(sub_cmd) => uplc::exec(sub_cmd),
+        Cmd::Completion(sub_cmd) => completion::exec(sub_cmd)
     }
-}
-
-fn version() -> String {
-    format!(
-        "v{} {}",
-        built_info::PKG_VERSION,
-        built_info::GIT_COMMIT_HASH_SHORT.unwrap_or("unknown")
-    )
 }
 
 fn panic_handler() {
